// src/app/interface/usuarios.interface.ts
export interface usuarios{
  id: number;
  usuario: string;
  contrasena: string;
  nombre: string;
  apellido: string;
  fecha_nacimiento: Date;
  altura: number;
  pesoActual: number;
  pesoObjetivo: number;
  nivel: string;
  rol: string;
  numero_tarjeta: string;
  nombre_titular: string;
  cvv: string;
  fecha_vencimiento: Date;
}